+++ src/bgp/ripc.c
@@ -46,11 +46,11 @@ ripc_send_short(
 		uint16_t src,
 		uint16_t dest,
 		void **buf,
-		size_t *length,
-		uint32_t num_items,
+		uint32_t *length,
+		uint16_t num_items,
 		void **return_bufs,
-		size_t *return_buf_lengths,
-		uint32_t num_return_bufs) {
+		uint32_t *return_buf_lengths,
+		uint16_t num_return_bufs) {
 
 	DEBUG("Starting short send: %u -> %u (%u items)", src, dest, num_items);
 	return ret;
@@ -61,11 +61,11 @@ ripc_send_long(
 		uint16_t src,
 		uint16_t dest,
 		void **buf,
-		size_t *length,
-		uint32_t num_items,
+		uint32_t *length,
+		uint16_t num_items,
 		void **return_bufs,
-		size_t *return_buf_lengths,
-		uint32_t num_return_bufs) {
+		uint32_t *return_buf_lengths,
+		uint16_t num_return_bufs) {
 
 	DEBUG("Starting long send: %u -> %u (%u items)", src, dest, num_items);
