+++ src/ripc.h
@@ -45,22 +45,22 @@ uint8_t ripc_send_short(
 		uint16_t src,
 		uint16_t dest,
 		void **buf,
-		size_t *length,
-		uint32_t num_items,
+		uint32_t *length,
+		uint16_t num_items,
 		void **return_bufs,
-		size_t *return_buf_lengths,
-		uint32_t num_return_bufs
+		uint32_t *return_buf_lengths,
+		uint16_t num_return_bufs
 		);
 
 uint8_t ripc_send_long(
 		uint16_t src,
 		uint16_t dest,
 		void **buf,
-		size_t *length,
-		uint32_t num_items,
+		uint32_t *length,
+		uint16_t num_items,
 		void **return_bufs,
-		size_t *return_buf_lengths,
-		uint32_t num_return_bufs
+		uint32_t *return_buf_lengths,
+		uint16_t num_return_bufs
 		);
 
 uint8_t ripc_receive(
